refactor(db): split base Prisma extension into smaller, descriptive modules
