+++ packages/db/prisma/extensions/index.ts
@@ -1,3 +1,5 @@
-export * from "./base";
 export * from "./custom-fields";
+export * from "./eth-usd-price";
+export * from "./helpers";
 export * from "./stats";
+export * from "./upsert-many";
